requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith>=0.2.0
langchain-community
langchain-core
langchain-google-genai
google-generativeai
notebook
python-dotenv
lxml
scikit-learn
pandas
pyarrow